utils: move run() here
